pyproject.toml
[tool.poetry]
name = "ArticleNoveltySearch"
version = "0.1.0"
description = ""
authors = ["Tesla2000 <fratajczak124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
psycopg2 = "^2.9"
pre-commit = "^3.7.1"
dynamic-executor = "^1.2.6"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
phidata = "^2.4.25"
sqlalchemy = "^2.0.31"
pgvector = "^0.3.1"
psycopg = "^3.2.1"
openai = "^1.35.13"
pypdf = "^4.2.0"
requests = "^2.32.3"
scikit-learn = "^1.5.1"
pyarrow = "^16.1.0"
arxiv = "^2.1.3"
more-itertools = "^10.3.0"
coverup = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
